+++ spec/ParseInstallation.spec.js
@@ -7,6 +7,7 @@ const Config = require('../lib/Config');
 const Parse = require('parse/node').Parse;
 const rest = require('../lib/rest');
 const request = require('../lib/request');
+const { ErrorMessage } = require('./Errors/message');
 
 let config;
 let database;
@@ -1232,7 +1233,7 @@ describe('Installations', () => {
           },
           err => {
             expect(err.code).toBe(136);
-            expect(err.message).toBe('installationId may not be changed in this operation');
+            expect(err.message).toBe(ErrorMessage.NO_OP('installationId'));
             done();
           }
         );
