+++ spec/rest.spec.js
@@ -6,6 +6,7 @@ const Parse = require('parse/node').Parse;
 const rest = require('../lib/rest');
 const RestWrite = require('../lib/RestWrite');
 const request = require('../lib/request');
+const { ErrorMessage } = require('./Errors/message');
 
 let config;
 let database;
@@ -77,7 +78,7 @@ describe('rest create', () => {
       objectId: '',
     };
 
-    const err = 'objectId must not be empty, null or undefined';
+    const err = ErrorMessage.MISSING_OBJECT_ID();
 
     expect(() => rest.create(config, auth.nobody(config), 'MyClass', objIdEmpty)).toThrowError(err);
