+++ src/Errors/message.js
@@ -6,11 +6,11 @@ export const ErrorMessage = {
     'This user is not allowed to access ' + 'non-existent class: ' + className,
   OBJECT_NOT_FOUND: () => 'Object not found.',
   FIELD_IS_REQUIRED: fieldName =>
-    'A value for the ' + fieldName + ' test is required to save/update the object.',
-  USERNAME_MISSING: () => 'bad or missing username',
-  PASSWORD_MISSING: () => 'password is required',
+    'A value for the ' + fieldName + ' is required to save/update the object.',
+  USERNAME_MISSING: () => 'Bad or missing username',
+  PASSWORD_MISSING: () => 'Password is required',
   UNSUPPORTED_SERVICE: () => 'This authentication method is unsupported.',
-  ACCOUNT_LINKED: () => 'this auth is already used',
+  ACCOUNT_LINKED: () => 'This auth is already used',
   CLIENT_EMAIL_VERIFICATION: () => "Clients aren't allowed to manually update email verification.",
   USERNAME_TAKEN: () => 'Account already exists for this username.',
   INVALID_EMAIL_ADDRESS_FORMAT: () => 'Email address format is invalid.',
@@ -23,10 +23,10 @@ export const ErrorMessage = {
   ACL_SESSION: () => 'Cannot set ' + 'ACL on a Session.',
   SESSION_ERROR: () => 'Error creating session.',
   MIN_ONE_ID_FIELD: () =>
-    'at least one ID field (deviceToken, installationId) must be specified in this operation',
+    'At least one ID field (deviceToken, installationId) must be specified in this operation',
   UNKNOWN_OBJECT_UPDATE: () => 'Object not found for update.',
-  NO_OP: field => field + ' may not be changed in this operation',
-  REQ_OP: field => field + ' must be specified in this operation',
+  NO_OP: field => 'The ' + field + ' can not be changed in this operation',
+  REQ_OP: field => 'The ' + field + ' must be specified in this operation',
   INSTALLATION_ID_WITH_DEVICE_TOKEN: () =>
     'Must specify installationId when deviceToken matches multiple Installation objects',
   USER_NO_MODIFY: field => 'Cannot modify user' + field + '.',
